requirements.txt
# Alpha Arena MVP Dependencies
openai>=1.0.0
anthropic>=0.7.0
requests>=2.28.0
python-dotenv>=1.0.0
